Usage of section templates instead of the project ones (if exist)
This commit provide a solution for the duplicate template element error
when saving a XSL template, or a XML document (in its XSL transformation
process).
At this point, if there is not a docxap.xsl file for the section, the
project main is used. Still if there are templates in a parent section.